Add TravisCI configuration

+++ generator/src/main/java/event/logging/gen/GenClasses.java
@@ -288,18 +288,10 @@ public class GenClasses {
         Files.walkFileTree(dir, new SimpleFileVisitor<Path>() {
             @Override
             public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                Files.delete(file);
-                return FileVisitResult.CONTINUE;
-            }
-            @Override
-            public FileVisitResult postVisitDirectory(Path dir, IOException e) throws IOException {
-                if (e == null) {
-                    Files.delete(dir);
-                    return FileVisitResult.CONTINUE;
-                } else {
-                    // directory iteration failed
-                    throw e;
+                if(!file.getFileName().toString().contains(".gitkeep")) {
+                    Files.delete(file);
                 }
+                return FileVisitResult.CONTINUE;
             }
         });
     }
